requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
python-dotenv
passlib[bcrypt]
slowapi
